+++ src/components/RoutesList.js
@@ -1,15 +1,15 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-import LineLink from './LineLink';
+import RouteLink from './RouteLink';
 
-class LinesList extends Component {
+class RoutesList extends Component {
   render() {
     return (
       <div className="overflow-scroll vh-75 ba ma2">
           {this.props.lines.map(line =>
             <div className="ph3 pv2 bb b--light-silver" key={line.id}>
-              <LineLink key={line.id} id={line.id} routeId={line.rt_id} name={line.rt_name} color={line.color} />
+              <RouteLink key={line.id} id={line.id} routeId={line.rt_id} name={line.rt_name} color={line.color} />
             </div>
           )}
       </div>
@@ -17,7 +17,7 @@ class LinesList extends Component {
   }
 }
 
-LinesList.propTypes = {
+RoutesList.propTypes = {
   lines: PropTypes.arrayOf(PropTypes.shape({
     agencyId: PropTypes.string,
     color: PropTypes.string,
@@ -29,4 +29,4 @@ LinesList.propTypes = {
   })).isRequired,
 }
 
-export default LinesList;
+export default RoutesList;
