+++ src/components/StopRouteList.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import _ from 'lodash'
 
 import Schedules from '../data/schedules.js'
-import LineLink from './LineLink'
+import RouteLink from './RouteLink'
 import RoutePredictionList from './RoutePredictionList'
 
 class StopRouteList extends Component {
@@ -15,7 +15,7 @@ class StopRouteList extends Component {
       <div className='list overflow-scroll'>
         {routes.map((r, i) => (
           <div className="pa2 bb" style={{ display: 'flex', alignItems: 'center' }} key={i}>
-            <LineLink name={Schedules[r].rt_name} id={r} color={Schedules[r].color} />
+            <RouteLink name={Schedules[r].rt_name} id={r} color={Schedules[r].color} />
             <RoutePredictionList predictions={_.filter(goodPredictions, function(o) { 
               return o.routeShortName === r.padStart(3, '0')})} route={r} />
           </div>
